+++ MathExpressionEvaluator/Expressions/Unary/Concrete/TangentUnaryExpression.cs
@@ -9,9 +9,9 @@ namespace MathExpressionEvaluator.Expressions.Unary.Concrete
         {
         }
 
-        internal override decimal Evaluate()
+        internal override double Evaluate()
         {
-            return (decimal) Math.Tan((double) Expression.Evaluate());
+            return Math.Tan(Expression.Evaluate());
         }
     }
 }
